pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
description = ""
authors = [
    {name = "kbs1115",email = "'bruce11158@gmail.com'"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.6,<0.116.0)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "python-jose (>=3.3.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "orjson (>=3.10.15,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
